pyproject.toml
[project]
name = "mcp-second-brain-server"
version = "0.1.0"
description = "Custom MCP server for Second Brain Obsidian vault with convention enforcement"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Jon", email = "jon@example.com" }
]

dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "fastapi[standard]>=0.115.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
    "qdrant-client>=1.7.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "aiofiles>=24.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.6",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vault", "src/models.py"]
only-include = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
addopts = "--cov=src --cov-report term-missing"
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/", "build/"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]
